Split controllers into API and GUI controllers
