+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/TopicsAPIController.java
@@ -1,0 +1,135 @@
+package de.digitalcollections.cudami.admin.controller.identifiable.entity;
+
+import de.digitalcollections.cudami.admin.controller.AbstractPagingAndSortingController;
+import de.digitalcollections.cudami.admin.util.LanguageSortingHelper;
+import de.digitalcollections.cudami.client.CudamiClient;
+import de.digitalcollections.cudami.client.CudamiLocalesClient;
+import de.digitalcollections.cudami.client.identifiable.entity.CudamiTopicsClient;
+import de.digitalcollections.model.exception.TechnicalException;
+import de.digitalcollections.model.identifiable.entity.Entity;
+import de.digitalcollections.model.identifiable.entity.Topic;
+import de.digitalcollections.model.identifiable.resource.FileResource;
+import de.digitalcollections.model.list.paging.PageRequest;
+import de.digitalcollections.model.list.paging.PageResponse;
+import de.digitalcollections.model.list.sorting.Order;
+import de.digitalcollections.model.list.sorting.Sorting;
+import java.util.List;
+import java.util.UUID;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+/** Controller for all public "Topics" endpoints (API). */
+@RestController
+public class TopicsAPIController extends AbstractPagingAndSortingController<Topic> {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(TopicsAPIController.class);
+
+  private final LanguageSortingHelper languageSortingHelper;
+  private final CudamiLocalesClient localeService;
+  private final CudamiTopicsClient service;
+
+  public TopicsAPIController(LanguageSortingHelper languageSortingHelper, CudamiClient client) {
+    this.languageSortingHelper = languageSortingHelper;
+    this.localeService = client.forLocales();
+    this.service = client.forTopics();
+  }
+
+  @GetMapping("/api/topics/new")
+  @ResponseBody
+  public Topic create() throws TechnicalException {
+    return service.create();
+  }
+
+  @GetMapping("/api/topics")
+  @ResponseBody
+  public PageResponse<Topic> findTop(
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
+      @RequestParam(name = "searchTerm", required = false) String searchTerm,
+      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
+      throws TechnicalException {
+    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
+    if (sortBy != null) {
+      Sorting sorting = new Sorting(sortBy);
+      pageRequest.setSorting(sorting);
+    }
+    return this.service.findTopTopics(pageRequest);
+  }
+
+  @GetMapping("/api/topics/{uuid}/topics")
+  @ResponseBody
+  public PageResponse<Topic> findSubtopic(
+      @PathVariable UUID uuid,
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
+      @RequestParam(name = "searchTerm", required = false) String searchTerm)
+      throws TechnicalException {
+    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
+    return service.findSubtopics(uuid, pageRequest);
+  }
+
+  @GetMapping("/api/topics/{uuid}/entities")
+  @ResponseBody
+  public PageResponse<Entity> getAttachedEntites(
+      @PathVariable UUID uuid,
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize)
+      throws TechnicalException {
+    return this.service.findEntities(uuid, new PageRequest(pageNumber, pageSize));
+  }
+
+  @GetMapping("/api/topics/{uuid}")
+  @ResponseBody
+  public Topic getByUuid(@PathVariable UUID uuid) throws TechnicalException {
+    return service.getByUuid(uuid);
+  }
+
+  @GetMapping("/api/topics/{uuid}/fileresources")
+  @ResponseBody
+  public PageResponse<FileResource> getRelatedFileResources(
+      @PathVariable UUID uuid,
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize)
+      throws TechnicalException {
+    return this.service.findFileResources(uuid, new PageRequest(pageNumber, pageSize));
+  }
+
+  @PostMapping("/api/topics")
+  public ResponseEntity save(
+      @RequestBody Topic topic,
+      @RequestParam(name = "parentUuid", required = false) UUID parentUuid) {
+    try {
+      Topic topicDb = null;
+      if (parentUuid == null) {
+        topicDb = service.save(topic);
+      } else {
+        topicDb = service.saveWithParentTopic(topic, parentUuid);
+      }
+      return ResponseEntity.status(HttpStatus.CREATED).body(topicDb);
+    } catch (TechnicalException e) {
+      LOGGER.error("Cannot save topic: ", e);
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+    }
+  }
+
+  @PutMapping("/api/topics/{uuid}")
+  public ResponseEntity update(@PathVariable UUID uuid, @RequestBody Topic topic) {
+    try {
+      Topic topicDb = service.update(uuid, topic);
+      return ResponseEntity.ok(topicDb);
+    } catch (TechnicalException e) {
+      LOGGER.error("Cannot save topic with uuid={}", uuid, e);
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+    }
+  }
+}
