+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/WebsitesController.java
@@ -8,11 +8,6 @@ import de.digitalcollections.cudami.client.identifiable.entity.CudamiWebsitesCli
 import de.digitalcollections.model.exception.ResourceNotFoundException;
 import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.identifiable.entity.Website;
-import de.digitalcollections.model.identifiable.web.Webpage;
-import de.digitalcollections.model.list.paging.PageRequest;
-import de.digitalcollections.model.list.paging.PageResponse;
-import de.digitalcollections.model.list.sorting.Order;
-import de.digitalcollections.model.list.sorting.Sorting;
 import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
@@ -20,18 +15,12 @@ import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 /** Controller for website management pages. */
 @Controller
@@ -53,12 +42,6 @@ public class WebsitesController extends AbstractController {
   public String create(Model model) throws TechnicalException {
     model.addAttribute("activeLanguage", localeService.getDefaultLanguage());
     return "websites/create";
-  }
-
-  @GetMapping("/api/websites/new")
-  @ResponseBody
-  public Website create() throws TechnicalException {
-    return service.create();
   }
 
   @GetMapping("/websites/{uuid}/edit")
@@ -84,40 +67,6 @@ public class WebsitesController extends AbstractController {
     return "websites/edit";
   }
 
-  @GetMapping("/api/websites")
-  @ResponseBody
-  public PageResponse<Website> find(
-      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
-      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
-      @RequestParam(name = "searchTerm", required = false) String searchTerm,
-      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
-      throws TechnicalException {
-    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
-    if (sortBy != null) {
-      Sorting sorting = new Sorting(sortBy);
-      pageRequest.setSorting(sorting);
-    }
-    return service.find(pageRequest);
-  }
-
-  @GetMapping("/api/websites/{uuid}/webpages")
-  @ResponseBody
-  public PageResponse<Webpage> findRootpages(
-      @PathVariable UUID uuid,
-      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
-      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
-      @RequestParam(name = "searchTerm", required = false) String searchTerm)
-      throws TechnicalException {
-    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
-    return service.findRootWebpages(uuid, pageRequest);
-  }
-
-  @GetMapping("/api/websites/{uuid}")
-  @ResponseBody
-  public Website getByUuid(@PathVariable UUID uuid) throws TechnicalException {
-    return service.getByUuid(uuid);
-  }
-
   @GetMapping("/websites")
   public String list(Model model) throws TechnicalException {
     final Locale displayLocale = LocaleContextHolder.getLocale();
@@ -130,38 +79,6 @@ public class WebsitesController extends AbstractController {
   @ModelAttribute("menu")
   protected String module() {
     return "websites";
-  }
-
-  @PostMapping("/api/websites")
-  public ResponseEntity save(@RequestBody Website website) {
-    try {
-      Website websiteDb = service.save(website);
-      return ResponseEntity.status(HttpStatus.CREATED).body(websiteDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save website: ", e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
-  }
-
-  @PutMapping("/api/websites/{uuid}")
-  public ResponseEntity update(@PathVariable UUID uuid, @RequestBody Website website) {
-    try {
-      Website websiteDb = service.update(uuid, website);
-      return ResponseEntity.ok(websiteDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save website with uuid={}", uuid, e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
-  }
-
-  @PutMapping("/api/websites/{uuid}/webpages")
-  public ResponseEntity updateRootPagesOrder(
-      @PathVariable UUID uuid, @RequestBody List<Webpage> rootPages) throws TechnicalException {
-    boolean successful = service.updateRootWebpagesOrder(uuid, rootPages);
-    if (successful) {
-      return new ResponseEntity<>(successful, HttpStatus.OK);
-    }
-    return new ResponseEntity<>(successful, HttpStatus.NOT_FOUND);
   }
 
   @GetMapping("/websites/{uuid}")
