+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/agent/PersonsAPIController.java
@@ -1,0 +1,96 @@
+package de.digitalcollections.cudami.admin.controller.identifiable.entity.agent;
+
+import de.digitalcollections.cudami.admin.controller.AbstractPagingAndSortingController;
+import de.digitalcollections.cudami.admin.util.LanguageSortingHelper;
+import de.digitalcollections.cudami.client.CudamiClient;
+import de.digitalcollections.cudami.client.CudamiLocalesClient;
+import de.digitalcollections.cudami.client.identifiable.entity.agent.CudamiPersonsClient;
+import de.digitalcollections.model.exception.TechnicalException;
+import de.digitalcollections.model.identifiable.entity.agent.Person;
+import de.digitalcollections.model.list.paging.PageRequest;
+import de.digitalcollections.model.list.paging.PageResponse;
+import de.digitalcollections.model.list.sorting.Order;
+import de.digitalcollections.model.list.sorting.Sorting;
+import java.util.List;
+import java.util.UUID;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+/** Controller for all public "Persons" endpoints (API). */
+@RestController
+public class PersonsAPIController extends AbstractPagingAndSortingController<Person> {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(PersonsAPIController.class);
+
+  private final LanguageSortingHelper languageSortingHelper;
+  private final CudamiLocalesClient localeService;
+  private final CudamiPersonsClient service;
+
+  @Autowired
+  public PersonsAPIController(LanguageSortingHelper languageSortingHelper, CudamiClient client) {
+    this.languageSortingHelper = languageSortingHelper;
+    this.localeService = client.forLocales();
+    this.service = client.forPersons();
+  }
+
+  @GetMapping("/api/persons/new")
+  @ResponseBody
+  public Person create() throws TechnicalException {
+    return service.create();
+  }
+
+  @GetMapping("/api/persons")
+  @ResponseBody
+  public PageResponse<Person> find(
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
+      @RequestParam(name = "searchTerm", required = false) String searchTerm,
+      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
+      throws TechnicalException {
+    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
+    if (sortBy != null) {
+      Sorting sorting = new Sorting(sortBy);
+      pageRequest.setSorting(sorting);
+    }
+    return service.find(pageRequest);
+  }
+
+  @GetMapping("/api/persons/{uuid}")
+  @ResponseBody
+  public Person getByUuid(@PathVariable UUID uuid) throws TechnicalException {
+    return service.getByUuid(uuid);
+  }
+
+  @PostMapping("/api/persons")
+  public ResponseEntity save(@RequestBody Person person) {
+    try {
+      Person personDb = service.save(person);
+      return ResponseEntity.status(HttpStatus.CREATED).body(personDb);
+    } catch (TechnicalException e) {
+      LOGGER.error("Cannot save person: ", e);
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+    }
+  }
+
+  @PutMapping("/api/persons/{uuid}")
+  public ResponseEntity update(@PathVariable UUID uuid, @RequestBody Person person) {
+    try {
+      Person personDb = service.update(uuid, person);
+      return ResponseEntity.ok(personDb);
+    } catch (TechnicalException e) {
+      LOGGER.error("Cannot save person with uuid={}", uuid, e);
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+    }
+  }
+}
