+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/agent/PersonsController.java
@@ -8,10 +8,6 @@ import de.digitalcollections.cudami.client.identifiable.entity.agent.CudamiPerso
 import de.digitalcollections.model.exception.ResourceNotFoundException;
 import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.identifiable.entity.agent.Person;
-import de.digitalcollections.model.list.paging.PageRequest;
-import de.digitalcollections.model.list.paging.PageResponse;
-import de.digitalcollections.model.list.sorting.Order;
-import de.digitalcollections.model.list.sorting.Sorting;
 import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
@@ -19,18 +15,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 /** Controller for Person management pages. */
 @Controller
@@ -55,12 +45,6 @@ public class PersonsController extends AbstractController {
     return "persons/create";
   }
 
-  @GetMapping("/api/persons/new")
-  @ResponseBody
-  public Person create() throws TechnicalException {
-    return service.create();
-  }
-
   @GetMapping("/persons/{uuid}/edit")
   public String edit(
       @PathVariable UUID uuid,
@@ -83,28 +67,6 @@ public class PersonsController extends AbstractController {
     return "persons/edit";
   }
 
-  @GetMapping("/api/persons")
-  @ResponseBody
-  public PageResponse<Person> find(
-      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
-      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
-      @RequestParam(name = "searchTerm", required = false) String searchTerm,
-      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
-      throws TechnicalException {
-    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
-    if (sortBy != null) {
-      Sorting sorting = new Sorting(sortBy);
-      pageRequest.setSorting(sorting);
-    }
-    return service.find(pageRequest);
-  }
-
-  @GetMapping("/api/persons/{uuid}")
-  @ResponseBody
-  public Person getByUuid(@PathVariable UUID uuid) throws TechnicalException {
-    return service.getByUuid(uuid);
-  }
-
   @GetMapping("/persons")
   public String list(Model model) throws TechnicalException {
     final Locale displayLocale = LocaleContextHolder.getLocale();
@@ -117,28 +79,6 @@ public class PersonsController extends AbstractController {
   @ModelAttribute("menu")
   protected String module() {
     return "persons";
-  }
-
-  @PostMapping("/api/persons")
-  public ResponseEntity save(@RequestBody Person person) {
-    try {
-      Person personDb = service.save(person);
-      return ResponseEntity.status(HttpStatus.CREATED).body(personDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save person: ", e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
-  }
-
-  @PutMapping("/api/persons/{uuid}")
-  public ResponseEntity update(@PathVariable UUID uuid, @RequestBody Person person) {
-    try {
-      Person personDb = service.update(uuid, person);
-      return ResponseEntity.ok(personDb);
-    } catch (TechnicalException e) {
-      LOGGER.error("Cannot save person with uuid={}", uuid, e);
-      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-    }
   }
 
   @GetMapping("/persons/{uuid}")
