+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/geo/location/GeoLocationsAPIController.java
@@ -1,0 +1,57 @@
+package de.digitalcollections.cudami.admin.controller.identifiable.entity.geo.location;
+
+import de.digitalcollections.commons.springmvc.controller.AbstractController;
+import de.digitalcollections.cudami.admin.util.LanguageSortingHelper;
+import de.digitalcollections.cudami.client.CudamiClient;
+import de.digitalcollections.cudami.client.identifiable.entity.geo.location.CudamiGeoLocationsClient;
+import de.digitalcollections.model.exception.TechnicalException;
+import de.digitalcollections.model.identifiable.entity.geo.location.GeoLocation;
+import de.digitalcollections.model.list.paging.PageRequest;
+import de.digitalcollections.model.list.paging.PageResponse;
+import de.digitalcollections.model.list.sorting.Order;
+import de.digitalcollections.model.list.sorting.Sorting;
+import java.util.List;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+/** Controller for all public "GeoLocations" endpoints (API). */
+@RestController
+public class GeoLocationsAPIController extends AbstractController {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(GeoLocationsAPIController.class);
+
+  private final LanguageSortingHelper languageSortingHelper;
+  private final CudamiGeoLocationsClient service;
+
+  public GeoLocationsAPIController(
+      LanguageSortingHelper languageSortingHelper, CudamiClient client) {
+    this.languageSortingHelper = languageSortingHelper;
+    this.service = client.forGeoLocations();
+  }
+
+  @GetMapping("/api/geolocations/new")
+  @ResponseBody
+  public GeoLocation create() throws TechnicalException {
+    return service.create();
+  }
+
+  @GetMapping("/api/geolocations")
+  @ResponseBody
+  public PageResponse<GeoLocation> find(
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
+      @RequestParam(name = "searchTerm", required = false) String searchTerm,
+      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
+      throws TechnicalException {
+    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
+    if (sortBy != null) {
+      Sorting sorting = new Sorting(sortBy);
+      pageRequest.setSorting(sorting);
+    }
+    return service.find(pageRequest);
+  }
+}
