+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/geo/location/GeoLocationsController.java
@@ -7,10 +7,6 @@ import de.digitalcollections.cudami.client.identifiable.entity.geo.location.Cuda
 import de.digitalcollections.model.exception.ResourceNotFoundException;
 import de.digitalcollections.model.exception.TechnicalException;
 import de.digitalcollections.model.identifiable.entity.geo.location.GeoLocation;
-import de.digitalcollections.model.list.paging.PageRequest;
-import de.digitalcollections.model.list.paging.PageResponse;
-import de.digitalcollections.model.list.sorting.Order;
-import de.digitalcollections.model.list.sorting.Sorting;
 import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
@@ -22,8 +18,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 /** Controller for GeoLocations management pages. */
 @Controller
@@ -37,28 +31,6 @@ public class GeoLocationsController extends AbstractController {
   public GeoLocationsController(LanguageSortingHelper languageSortingHelper, CudamiClient client) {
     this.languageSortingHelper = languageSortingHelper;
     this.service = client.forGeoLocations();
-  }
-
-  @GetMapping("/api/geolocations/new")
-  @ResponseBody
-  public GeoLocation create() throws TechnicalException {
-    return service.create();
-  }
-
-  @GetMapping("/api/geolocations")
-  @ResponseBody
-  public PageResponse<GeoLocation> find(
-      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
-      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
-      @RequestParam(name = "searchTerm", required = false) String searchTerm,
-      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
-      throws TechnicalException {
-    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
-    if (sortBy != null) {
-      Sorting sorting = new Sorting(sortBy);
-      pageRequest.setSorting(sorting);
-    }
-    return service.find(pageRequest);
   }
 
   @GetMapping("/geolocations")
