+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/legal/LicensesAPIController.java
@@ -1,0 +1,94 @@
+package de.digitalcollections.cudami.admin.controller.legal;
+
+import de.digitalcollections.cudami.admin.controller.AbstractPagingAndSortingController;
+import de.digitalcollections.cudami.admin.util.LanguageSortingHelper;
+import de.digitalcollections.cudami.client.CudamiClient;
+import de.digitalcollections.cudami.client.CudamiLocalesClient;
+import de.digitalcollections.cudami.client.legal.CudamiLicensesClient;
+import de.digitalcollections.model.exception.TechnicalException;
+import de.digitalcollections.model.legal.License;
+import de.digitalcollections.model.list.paging.PageRequest;
+import de.digitalcollections.model.list.paging.PageResponse;
+import de.digitalcollections.model.list.sorting.Order;
+import de.digitalcollections.model.list.sorting.Sorting;
+import java.util.List;
+import java.util.UUID;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+/** Controller for all public "Licenses" endpoints (API). */
+@RestController
+public class LicensesAPIController extends AbstractPagingAndSortingController<License> {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(LicensesAPIController.class);
+
+  private final LanguageSortingHelper languageSortingHelper;
+  private final CudamiLocalesClient localeService;
+  private final CudamiLicensesClient service;
+
+  public LicensesAPIController(LanguageSortingHelper languageSortingHelper, CudamiClient client) {
+    this.languageSortingHelper = languageSortingHelper;
+    this.localeService = client.forLocales();
+    this.service = client.forLicenses();
+  }
+
+  @GetMapping("/api/licenses/new")
+  @ResponseBody
+  public License createModel() throws TechnicalException {
+    return service.create();
+  }
+
+  @GetMapping("/api/licenses")
+  @ResponseBody
+  public PageResponse<License> find(
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "25") int pageSize,
+      @RequestParam(name = "searchTerm", required = false) String searchTerm,
+      @RequestParam(name = "sortBy", required = false) List<Order> sortBy)
+      throws TechnicalException {
+    PageRequest pageRequest = new PageRequest(searchTerm, pageNumber, pageSize);
+    if (sortBy != null) {
+      Sorting sorting = new Sorting(sortBy);
+      pageRequest.setSorting(sorting);
+    }
+    return service.find(pageRequest);
+  }
+
+  @GetMapping("/api/licenses/{uuid}")
+  @ResponseBody
+  public License getByUuid(@PathVariable UUID uuid) throws TechnicalException {
+    return service.getByUuid(uuid);
+  }
+
+  @PostMapping("/api/licenses")
+  public ResponseEntity save(@RequestBody License license) {
+    try {
+      License licenseDb = service.save(license);
+      return ResponseEntity.status(HttpStatus.CREATED).body(licenseDb);
+    } catch (TechnicalException e) {
+      LOGGER.error("Cannot save license: ", e);
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+    }
+  }
+
+  @PutMapping("/api/licenses/{uuid}")
+  public ResponseEntity update(@PathVariable UUID uuid, @RequestBody License license) {
+    try {
+      License licenseDb = service.update(uuid, license);
+      return ResponseEntity.ok(licenseDb);
+    } catch (TechnicalException e) {
+      LOGGER.error("Cannot update license with uuid={}", uuid, e);
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+    }
+  }
+}
